+++ consensus/common/src/main/java/org/hyperledger/besu/consensus/common/bft/blockcreation/BftBlockCreator.java
@@ -68,7 +68,7 @@ public class BftBlockCreator extends AbstractBlockCreator {
       final EthScheduler ethScheduler) {
     super(
         miningConfiguration.setCoinbase(localAddress),
-        miningBeneficiaryCalculator(localAddress, forksSchedule),
+        miningBeneficiaryCalculator(localAddress, protocolSchedule),
         extraDataCalculator,
         transactionPool,
         protocolContext,
@@ -91,9 +91,19 @@ public class BftBlockCreator extends AbstractBlockCreator {
   }
 
   private static MiningBeneficiaryCalculator miningBeneficiaryCalculator(
-      final Address localAddress, final ForksSchedule<? extends BftConfigOptions> forksSchedule) {
-    return blockNum ->
-        forksSchedule.getFork(blockNum).getValue().getMiningBeneficiary().orElse(localAddress);
+      final Address localAddress, final ProtocolSchedule protocolSchedule) {
+    return (blockTimestamp, pendingHeader) -> {
+      BlockHeader newBlockHeader =
+          BlockHeaderBuilder.createDefault()
+              .coinbase(pendingHeader.getCoinbase())
+              .buildBlockHeader();
+      ProtocolSpec protocolSpec =
+          ((BftProtocolSchedule) protocolSchedule)
+              .getByBlockNumberOrTimestamp(pendingHeader.getNumber(), blockTimestamp);
+      Address beneficiaryAddress =
+          protocolSpec.getMiningBeneficiaryCalculator().calculateBeneficiary(newBlockHeader);
+      return !beneficiaryAddress.isZero() ? beneficiaryAddress : localAddress;
+    };
   }
 
   @Override